unit tests passing for ingest
